requirements.txt
apturl==0.5.2
asgiref==3.7.2
attrs==21.2.0
axios==0.4.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
et-xmlfile==1.1.0
fasteners==0.14.1
filelock==3.13.1
future==0.18.2
gunicorn==21.2.0
gyp==0.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
infobip-api-python-client==3.0.3
infobip-api-python-sdk==5.0.1
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mysqlclient==2.2.1
netifaces==0.11.0
numpy==1.26.3
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
paramiko==2.9.3
pexpect==4.8.0
pillow==10.2.0
platformdirs==4.1.0
protobuf==3.12.4
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycryptodome==3.20.0
pycups==2.0.1
pydantic==1.10.13
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pypng==0.20220715.0
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-barcode==0.15.1
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.4.2
regex==2022.3.15
reportlab==4.0.7
requests==2.31.0
rest-framework-simplejwt==0.0.2
rich==13.7.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
texttable==1.6.4
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.1.0
usb-creator==0.3.7
virtualenv==20.25.0
wadllib==1.3.6
websocket-client==1.2.3
whitenoise==6.6.0
xdg==5
xkit==0.0.0
zipp==1.0.0
